feat: create formation by category and get all ..

+++ src/app/pages/gestion-formation/gestion-formation.component.ts
@@ -1,9 +1,11 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { FormationService } from 'src/app/_services/formation/formation.service';
 import { StorageService } from 'src/app/_services/storage.service';
 import { FormationModalComponent } from '../formation-modal/formation-modal.component';
 import { FormationCategory } from 'src/app/pages/gestion-formation/formation-category.enum';
+import { MatPaginator } from '@angular/material/paginator';
+import { MatTableDataSource } from '@angular/material/table';
 
 @Component({
   selector: 'app-gestion-formation',
@@ -12,29 +14,32 @@ import { FormationCategory } from 'src/app/pages/gestion-formation/formation-cat
 })
 export class GestionFormationComponent implements OnInit {
   formations: any[] = [];
-  currentPageformations: any[] = [];
-  currentPage: number = 1;
-  itemsPerPage: number = 5;
-  totalPages: number = 0;
-
+  dataSource: MatTableDataSource<any> = new MatTableDataSource<any>([]);
+  displayedColumns: string[] = ['title', 'description', 'location', 'startDate', 'endDate', 'price', 'numberOfHours', 'category', 'bestSeller', 'newFormation', 'actions'];
   nouvellesformations: number = 0;
   formationsExpirees: number = 0;
   categories: string[] = Object.values(FormationCategory);
-  userId:any
-  constructor(private formationService: FormationService, private modalService: NgbModal,private storageService:StorageService) { }
+  userId: any;
+
+  @ViewChild(MatPaginator) paginator: MatPaginator;
+
+  constructor(private formationService: FormationService, private modalService: NgbModal, private storageService: StorageService) { }
 
   ngOnInit(): void {
     this.loadAllformations();
-    
-   this.userId= this.storageService.getUser()
+    this.userId = this.storageService.getUser();
+  }
+
+  ngAfterViewInit() {
+    this.dataSource.paginator = this.paginator;
   }
 
   loadAllformations(): void {
     this.formationService.getAllFormations().subscribe(
       (data: any[]) => {
         this.formations = data;
+        this.dataSource.data = data; // Mettez à jour le dataSource pour le tableau Material
         this.calculateStatistics();
-        this.updateCurrentPageformations();
       },
       error => {
         console.error('Erreur lors de la récupération des formations :', error);
@@ -56,56 +61,33 @@ export class GestionFormationComponent implements OnInit {
     }).length;
   }
 
-  updateCurrentPageformations(): void {
-    this.totalPages = Math.ceil(this.formations.length / this.itemsPerPage);
-    const startIndex = (this.currentPage - 1) * this.itemsPerPage;
-    const endIndex = startIndex + this.itemsPerPage;
-    this.currentPageformations = this.formations.slice(startIndex, endIndex);
-  }
-
-  changePage(page: number): void {
-    if (page > 0 && page <= this.totalPages) {
-      this.currentPage = page;
-      this.updateCurrentPageformations();
-    }
-  }
-  deleteFormation(id: number): void {
-    this.formationService.deleteFormation(id).subscribe(() => {
-      this.formations = this.formations.filter(formation => formation.id !== id);
-      this.calculateStatistics();
-      this.updateCurrentPageformations();
-    }, error => {
-      console.error('Erreur lors de la suppression de l\'formation :', error);
-    });
-  }
-
   openModal(isEditing: boolean, formation?: any): void {
     const modalRef = this.modalService.open(FormationModalComponent);
-    modalRef.componentInstance.formation = formation || {}; // Initialiser l'formation avec celle fournie ou un objet vide
+    modalRef.componentInstance.formation = formation || {}; 
     modalRef.componentInstance.isEditing = isEditing;
 
     modalRef.componentInstance.save.subscribe((result: any) => {
-      //result.user=this.userId
-       console.log(result)
       this.formationService.createFormation(result).subscribe((newformation: any) => {
         console.log('formation ajoutée avec succès', newformation);
         this.formations.push(newformation);
-      })
-      modalRef.close();
+        this.dataSource.data = this.formations; // Mettre à jour le dataSource
+        modalRef.close();
+      });
     });
 
     modalRef.componentInstance.cancel.subscribe(() => {
-      modalRef.close(); // Fermer la modal en cas d'annulation
+      modalRef.close(); 
     });
   }
 
-  addFormation(formation: any): void {
-    this.formationService.createFormation(formation).subscribe((newformation: any) => {
-      this.formations.push(newformation);
+  deleteFormation(id: number): void {
+    if (confirm('Are you sure to delete this formation ?'))
+    this.formationService.deleteFormation(id).subscribe(() => {
+      this.formations = this.formations.filter(formation => formation.id !== id);
+      this.dataSource.data = this.formations; // Mettre à jour le dataSource
       this.calculateStatistics();
-      this.updateCurrentPageformations();
     }, error => {
-      console.error('Erreur lors de l\'ajout de l\'formation :', error);
+      console.error('Erreur lors de la suppression de l\'formation :', error);
     });
   }
 
@@ -114,9 +96,9 @@ export class GestionFormationComponent implements OnInit {
       const index = this.formations.findIndex(o => o.id === formation.id);
       if (index !== -1) {
         this.formations[index] = { ...formation };
+        this.dataSource.data = this.formations; // Mettre à jour le dataSource
       }
       this.calculateStatistics();
-      this.updateCurrentPageformations();
     }, error => {
       console.error('Erreur lors de la mise à jour de l\'formation :', error);
     });
